+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Ignore `Style/IndentationConsistency` RuboCop warnings
 * Add `ConsecutiveControlStatements` which recommends condensing multiple
   control statements into a single `ruby:` filter
+* Add `EmptyControlStatement` which reports control statements with no code
 
 ## 0.1.0